requirements.txt
PyYAML==6.0
jira==3.1.1
python-dotenv==0.21.0
pandas==2.0.3
langchain==0.1.4
langchain-community==0.0.20
langchain-google-vertexai==0.0.2
SQLAlchemy==2.0.43
